Handle M-A:same conflicts with virtual packages
We have a Multi-Arch: Same package A whose Conflicts matches itself, either the
real package or a virtual provides. For non-virtual packages, things are easy,
we can just drop the Conflicts, since self-conflicts are supposed to be
ignored, and this avoids the nastiness that's to follow.

However, if Conflicts is a virtual package (regardless of whether it is also a
real package), we need to be very careful. We want to conflict with every other
provider of that virtual package, but not ourselves, not even other
architectures, which means we cannot conflict with --virtual-foo:arch (cudf
will allow A:arch1 to conflict with A:arch1, but not A:arch2).  Thus we must
expand out any self-conflicting virtual Conflicts to the real packages
providing them, with extra care taken in case the package name is both real and
virtual (we would otherwise get the desired foo:arch with an undesired
--virtual-foo:arch, the very thing we are trying to get rid of!). Thus we make
sure to call add_arch_info manually here rather than loadl on the whole result
in order to bypass the virtual_table check.